PV Today

+++ src/Inverters/InverterResult.hpp
@@ -40,7 +40,7 @@ typedef struct
     int inverterTemperature = 0;
     int batteryTemperature = 0;
     double pvToday = 0;
-    uint32_t pvTotal = 0;
+    double pvTotal = 0;
     bool hasBattery = true;
 } InverterData_t;
